Create gfs_read.py
xarray uses cfgrib to read the grib or grib2 files. cfgrib have to explicitly give in what type of level you are reading the data, especially in the case of NCEP GFS data. GFS contains data in variety of types of levels.   Details are given in the read me file of this git.